Aggiunto gestione segnali quando il parent process esce.

+++ lib/include/arguments.h
@@ -4,9 +4,9 @@
 #include "./list.h"
 
 typedef struct FarmArguments {
-    int     farm_setup_threads_number;
-    int     farm_setup_queue_length;
-    int     farm_setup_delay_time;
+    long     farm_setup_threads_number;
+    long     farm_setup_queue_length;
+    long    farm_setup_delay_time;
     char *  farm_setup_directory_name;
     list *  farm_setup_file_list;
 } FarmArguments;
